Dedicated image variant dao
