+++ mdm/api/src/main/java/com/gentics/mesh/core/data/dao/BinaryDao.java
@@ -1,21 +1,15 @@
 package com.gentics.mesh.core.data.dao;
 
 import java.io.InputStream;
-import java.util.Collection;
 import java.util.stream.Stream;
 
-import com.gentics.mesh.context.InternalActionContext;
 import com.gentics.mesh.core.data.HibBinaryDataElement;
 import com.gentics.mesh.core.data.binary.HibBinary;
-import com.gentics.mesh.core.data.binary.HibImageVariant;
 import com.gentics.mesh.core.data.node.field.HibBinaryField;
 import com.gentics.mesh.core.db.Supplier;
 import com.gentics.mesh.core.db.Transactional;
 import com.gentics.mesh.core.rest.node.field.BinaryCheckStatus;
-import com.gentics.mesh.core.rest.node.field.image.ImageVariantRequest;
-import com.gentics.mesh.core.rest.node.field.image.ImageVariantResponse;
 import com.gentics.mesh.core.result.Result;
-import com.gentics.mesh.parameter.image.ImageManipulation;
 import com.gentics.mesh.util.UUIDUtil;
 
 import io.reactivex.Flowable;
@@ -24,7 +18,7 @@ import io.vertx.core.buffer.Buffer;
 /**
  * DAO for {@link HibBinary}.
  */
-public interface BinaryDao extends Dao<HibBinary>, DaoTransformable<HibImageVariant, ImageVariantResponse> {
+public interface BinaryDao extends Dao<HibBinary> {
 
 	/**
 	 * Return the binary data stream.
@@ -100,58 +94,4 @@ public interface BinaryDao extends Dao<HibBinary>, DaoTransformable<HibImageVari
 	 * @return
 	 */
 	Transactional<Stream<HibBinary>> findAll();
-
-	/**
-	 * Get the manipulation variants of this binary image representation. On any other binary data returns null.
-	 * 
-	 * @return
-	 */
-	Result<? extends HibImageVariant> getVariants(HibBinaryField binaryField, InternalActionContext ac);
-
-	/**
-	 * Get the manipulation variants of this binary image representation. On any other binary data returns null.
-	 * 
-	 * @return
-	 */
-	HibImageVariant getVariant(HibBinaryField binaryField, ImageManipulation variant, InternalActionContext ac);
-
-	/**
-	 * Create the image manipulation variant.
-	 * 
-	 * @param variant
-	 * @return created variants
-	 */
-	HibImageVariant createVariant(HibBinaryField binaryField, ImageVariantRequest variant, InternalActionContext ac, boolean throwOnExisting);
-
-	/**
-	 * Delete the image manipulation variant.
-	 * 
-	 * @param variant
-	 */
-	void deleteVariant(HibBinaryField binaryField, ImageVariantRequest variant, InternalActionContext ac, boolean throwOnAbsent);
-
-	/**
-	 * Create the image manipulation variants, optionally deleting all other existing unused variants.
-	 * 
-	 * @param variants
-	 * @param deleteOtherVariants
-	 * @return image variants of the binary after the operation.
-	 */
-	Result<? extends HibImageVariant> createVariants(HibBinaryField binaryField, Collection<ImageVariantRequest> variants, InternalActionContext ac, boolean deleteOtherVariants);
-
-	/**
-	 * Delete the requested image manipulation variants.
-	 * 
-	 * @param variant
-	 * @return image variants of the binary after the operation.
-	 */
-	Result<? extends HibImageVariant> deleteVariants(HibBinaryField binaryField, Collection<ImageVariantRequest> variant, InternalActionContext ac);
-
-	/**
-	 * Retain the requested image manipulation variants, delete all other.
-	 * 
-	 * @param variant
-	 * @return image variants of the binary after the operation.
-	 */
-	Result<? extends HibImageVariant> retainVariants(HibBinaryField binaryField, Collection<ImageVariantRequest> variant, InternalActionContext ac);
 }
