+++ mdm/api/src/main/java/com/gentics/mesh/core/data/dao/DaoCollection.java
@@ -18,6 +18,8 @@ import com.gentics.mesh.core.action.UserDAOActions;
  */
 public interface DaoCollection {
 
+	ImageVariantDao imageVariantDao();
+
 	UserDao userDao();
 
 	UserDAOActions userActions();
