+++ mdm/orientdb-api/src/main/java/com/gentics/mesh/core/data/dao/ImageVariantDaoWrapper.java
@@ -1,0 +1,11 @@
+package com.gentics.mesh.core.data.dao;
+
+/**
+ * OrientDB extension to {@link PersistingImageVariantDao}
+ * 
+ * @author plyhun
+ *
+ */
+public interface ImageVariantDaoWrapper extends PersistingImageVariantDao {
+
+}
