+++ mdm/orientdb-api/src/main/java/com/gentics/mesh/core/db/GraphDBTx.java
@@ -4,6 +4,7 @@ import com.gentics.mesh.core.data.dao.BinaryDaoWrapper;
 import com.gentics.mesh.core.data.dao.BranchDaoWrapper;
 import com.gentics.mesh.core.data.dao.ContentDaoWrapper;
 import com.gentics.mesh.core.data.dao.GroupDaoWrapper;
+import com.gentics.mesh.core.data.dao.ImageVariantDaoWrapper;
 import com.gentics.mesh.core.data.dao.JobDaoWrapper;
 import com.gentics.mesh.core.data.dao.LanguageDaoWrapper;
 import com.gentics.mesh.core.data.dao.MicroschemaDaoWrapper;
@@ -54,6 +55,9 @@ public interface GraphDBTx extends CommonTx, BaseTransaction {
 	static GraphDBTx getGraphTx() {
 		return (GraphDBTx) Tx.get();
 	}
+
+	@Override
+	ImageVariantDaoWrapper imageVariantDao();
 
 	@Override
 	NodeDaoWrapper nodeDao();
