+++ verticles/graphql/src/main/java/com/gentics/mesh/graphql/filter/BinaryFilter.java
@@ -1,18 +1,12 @@
 package com.gentics.mesh.graphql.filter;
 
-import java.util.Collection;
 import java.util.List;
 
 import com.gentics.graphqlfilter.filter.FilterField;
 import com.gentics.graphqlfilter.filter.MappedFilter;
 import com.gentics.graphqlfilter.filter.StringFilter;
-import com.gentics.graphqlfilter.filter.operation.JoinPart;
 import com.gentics.mesh.core.data.binary.HibBinary;
-import com.gentics.mesh.core.data.binary.HibImageVariant;
-import com.gentics.mesh.core.db.CommonTx;
 import com.gentics.mesh.graphql.context.GraphQLContext;
-
-import graphql.util.Pair;
 
 /**
  * Binary filter.
@@ -46,15 +40,14 @@ public class BinaryFilter extends ImageDataFilter<HibBinary> {
 		this.context = context;
 	}
 
-	@SuppressWarnings("unchecked")
 	@Override
 	public List<FilterField<HibBinary, ?>> getFilters() {
 		List<FilterField<HibBinary, ?>> filters = super.getFilters();
 		filters.add(new MappedFilter<>(owner, "checksum", "Filters by SHA512 checksum", StringFilter.filter(),
 				content -> content == null ? null : content.getSHA512Sum()));
-		filters.add(new MappedFilter<>(owner, "variants", "Filters by image variants", 
-				ListFilter.imageVariantListFilter(context, "BINARY"),
-				content -> content == null ? null : (Collection<HibImageVariant>) CommonTx.get().binaryDao().getVariants(content, context).list(), Pair.pair("variants", new JoinPart("IMAGEVARIANT", "value"))));
+//		filters.add(new MappedFilter<>(owner, "variants", "Filters by image variants", 
+//				ListFilter.imageVariantListFilter(context, "BINARY"),
+//				content -> content == null ? null : (Collection<HibImageVariant>) CommonTx.get().binaryDao().getVariants(content, context).list(), Pair.pair("variants", new JoinPart("IMAGEVARIANT", "value"))));
 		return filters;
 	}
 }
